requirements.txt
flask
requests
python-bale-bot